A suggestion for new makefile
This is just a suggestion, and I am very open to not using it.
This is a bit more manual than the last one, and I think this
is the most important pros and cons:

Pro:
Handles python as well as C++, and lets us use command line args
in the main.cpp file, so that it can for example computs one thing
without also computing a different unrelated solutiuon first.

Con:
Requires us adding some things to the makefile when new functionality
is implemented, instead of the makefile just running everything.
This makefile probably also has some issues where computations
are done unnecessarily before making the pdf. So, even though
a plot is made, for some reason it seems like make will sometimes
remake it if the plot is required.